Updated libs.versions.toml and corrected visuals update in editor EditorViewmodel editConfigs were not send to Visualizer viewmodel causing no updates, corrected that Using player mute and unmute apis rather than setting the volume in AudioFilePlayerImpl.kt onMuteDevice AudioInfoExtractor.kt separates the logic for extracting media info in PlayerFileProviderImpl.kt PlayerFileProviderImpl.kt now only handles the audio metadata from media store
