+++ data/player/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
 	implementation(libs.androidx.media3.common)
 	implementation(libs.androidx.media3.exoplayer)
 	implementation(libs.androidx.media3.session)
+	implementation(libs.androidx.media3.inspector)
 
 	// futures to coroutine
 	implementation(libs.androidx.concurrent.futures.ktx)
