+++ data/recordings/src/main/java/com/eva/recordings/di/RecordingsProviderModule.kt
@@ -8,6 +8,7 @@ import com.eva.datastore.domain.repository.RecorderAudioSettingsRepo
 import com.eva.datastore.domain.repository.RecorderFileSettingsRepo
 import com.eva.location.domain.repository.LocationAddressProvider
 import com.eva.recordings.data.RecordingWidgetInteractorImpl
+import com.eva.recordings.data.provider.AudioInfoExtractorImpl
 import com.eva.recordings.data.provider.PlayerFileProviderImpl
 import com.eva.recordings.data.provider.RecorderFileProviderImpl
 import com.eva.recordings.data.provider.RecordingSecondaryDataProviderImpl
@@ -15,6 +16,7 @@ import com.eva.recordings.data.provider.StorageInfoProviderImpl
 import com.eva.recordings.data.provider.TrashRecordingsProviderApi29Impl
 import com.eva.recordings.data.provider.TrashRecordingsProviderImpl
 import com.eva.recordings.data.provider.VoiceRecordingsProviderImpl
+import com.eva.recordings.domain.provider.AudioInfoExtractor
 import com.eva.recordings.domain.provider.PlayerFileProvider
 import com.eva.recordings.domain.provider.RecorderFileProvider
 import com.eva.recordings.domain.provider.RecordingsSecondaryDataProvider
@@ -32,6 +34,15 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 object RecordingsProviderModule {
+
+	@Provides
+	@Singleton
+	fun providesAudioExtractor(
+		@ApplicationContext context: Context,
+		locationProvider: LocationAddressProvider,
+		settings: RecorderAudioSettingsRepo,
+	): AudioInfoExtractor =
+		AudioInfoExtractorImpl(context, settings, locationProvider)
 
 	@Provides
 	@Singleton
@@ -66,9 +77,8 @@ object RecordingsProviderModule {
 	@Singleton
 	fun providesPlayerFileProvider(
 		@ApplicationContext context: Context,
-		locationProvider: LocationAddressProvider,
-		settings: RecorderAudioSettingsRepo,
-	): PlayerFileProvider = PlayerFileProviderImpl(context, settings, locationProvider)
+		extractor: AudioInfoExtractor,
+	): PlayerFileProvider = PlayerFileProviderImpl(context, extractor)
 
 
 	@Provides
