+++ data/recordings/src/main/java/com/eva/recordings/domain/provider/AudioInfoExtractor.kt
@@ -1,0 +1,8 @@
+package com.eva.recordings.domain.provider
+
+import com.eva.recordings.domain.models.MediaMetaDataInfo
+
+fun interface AudioInfoExtractor {
+
+	suspend fun extractMediaData(uri: String): MediaMetaDataInfo?
+}
