+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.13.1"
+agp = "8.13.2"
 concurrentFuturesKtx = "1.3.0"
 datastore = "1.2.0"
 coreSplashscreen = "1.2.0"
@@ -17,23 +17,23 @@ kotlinxCollectionsImmutable = "0.4.0"
 kotlinxDatetime = "0.7.1"
 kotlinxSerializationJson = "1.9.0"
 lifecycleRuntimeKtx = "2.10.0"
-activityCompose = "1.12.1"
-composeBom = "2025.12.00"
-ksp = "2.3.0"
+activityCompose = "1.12.2"
+composeBom = "2025.12.01"
+ksp = "2.3.4"
 hilt = "2.57.2"
-media3Common = "1.8.0"
+media3Common = "1.9.0"
 navigationCompose = "2.9.6"
 playServicesLocationVersion = "21.3.0"
 roomCompiler = "2.8.4"
 uiTextGoogleFonts = "1.10.0"
 workRuntimeKtxVersion = "2.11.0"
 hiltWork = "1.3.0"
-protobufJavalite = "4.33.1"
+protobufJavalite = "4.33.2"
 protobuf_version = "0.9.5"
 protobuf_gen_java_lite = "3.0.0"
 materialIconExtended = "1.7.8"
 moduleGrapher = "0.13.0"
-mockk = "1.14.6"
+mockk = "1.14.7"
 coroutines = "1.10.2"
 turbine_version = "1.2.1"
 compileSdk = "36"
@@ -64,6 +64,7 @@ androidx-media3-exoplayer = { module = "androidx.media3:media3-exoplayer", versi
 androidx-media3-session = { module = "androidx.media3:media3-session", version.ref = "media3Common" }
 androidx-media3-transformer = { module = "androidx.media3:media3-transformer", version.ref = "media3Common" }
 androidx-media3-effects = { module = "androidx.media3:media3-effect", version.ref = "media3Common" }
+androidx-media3-inspector = { module = "androidx.media3:media3-inspector", version.ref = "media3Common" }
 androidx-media3-extractor = { module = "androidx.media3:media3-extractor", version.ref = "media3Common" }
 androidx-media3-decoder = { module = "androidx.media3:media3-decoder", version.ref = "media3Common" }
 androidx-media3-ui = { module = "androidx.media3:media3-ui-compose-material3", version.ref = "media3Common" }
